Move visit-start up
